i forgor

+++ 2022/rust/src/day/xi.rs
@@ -64,18 +64,18 @@ impl<'a> Monkey<'a> {
     }
 }
 
-pub fn part_one(input: &str) -> usize {
-    monkey_business(parse_monkeys(input), 20, |worry| worry / 3)
+pub fn part_one(input: &str) -> isize {
+    monkey_business(parse_monkeys(input), 20, |worry| worry / 3) as isize
 }
 
-pub fn part_two(input: &str) -> usize {
+pub fn part_two(input: &str) -> isize {
     let monkeys: Vec<Monkey> = parse_monkeys(input);
 
     let divs = monkeys.iter().map(|m| m.div);
     let lcm = divs.clone().reduce(std::ops::Mul::mul).unwrap()
         / divs.clone().fold(0, |acc, d| acc.gcd(d));
 
-    monkey_business(monkeys, 10_000, |worry| worry % lcm)
+    monkey_business(monkeys, 10_000, |worry| worry % lcm) as isize
 }
 
 fn monkey_business(
